Move commented-out constructor to static factory
Static factories are a way of acquiring objects, much like a
constructor. While somewhat more verbose to write, static factories can
be more flexible than normal constructors; for example, if you want to
create a "default" AquaFish and randomly-colored AquaFish, you can't
use two constructors, as both would have a single Aquarium argument,
making things ambiguous when one of the two constructors is used.
Static factory methods are not required to have the same name as the
class, so arguments are not as big of an issue.