Rewrite array notes
More or less a complete rewrite, as the first version doesn't sound as
good as it did when I first wrote it.

The main aim was to make things clearer and to introduce things in a
more natural way. In addition, some explanations were omitted in favor
of examples, as the explanations could get a bit bland.